# contracts/openapi.v1.yaml
openapi: 3.1.0
info:
  title: TI-Toolbox job server (tit.server)
  version: "1.0.0"
  description: >-
    v1 contract for the v3 Electron GUI. Every path and schema from
    `contracts/openapi.v0.yaml` is carried unchanged (the real server's `--dump-openapi` output
    must be a superset of both; CI gate `dev/contracts_check.py`); v1 adds the job engine, catalog
    CRUD, ViewSpec viewers, jailed file access, validate/plan preflight, settings, and the
    `/ws/jobs` live-job stream. Auth is unchanged from v0: the launcher passes
    `?token=<TIT_SERVER_TOKEN>` exactly once; the server mints a random session id (never the
    token), keeps it in memory, sets it as an HttpOnly, SameSite=Strict `tit_session` cookie and
    redirects to `/`. A server restart forgets every session (401 -> back to the launcher). API
    routes accept either that cookie or `Authorization: Bearer <token>` (scripts). Unauthenticated
    requests to any route other than `/api/health` and `/auth/session` get 401.
    `POST /auth/logout` forgets the session and clears the cookie.

    Config request bodies (`config` fields below) are `dict[str, Any]`, validated server-side with
    `tit.config_io.deserialize_config`. Their real shapes are the `contracts/schema.json` `$defs`
    (`SimulationConfig`, `Montage`, `FlexConfig`, `ExConfig`, `MExConfig`, `AnalyzerConfig`,
    `PreprocessConfig`, `QSIPrepConfig`, `QSIReconConfig`, `GroupComparisonConfig`,
...
    today.

...

...

...
    difference is never reported as drift.
...
servers:
...
paths:
...
  /api/validate/{kind}:
    post:
      tags: [validate]
      summary: Field-level validation of a config, no side effects (dataclass __post_init__ errors)
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - in: path
          name: kind
          required: true
          schema: { $ref: "#/components/schemas/PipelineKind" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [config]
              properties:
                config:
                  $ref: "#/components/schemas/PipelineConfig"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidateResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/view/{kind}:
    get:
      tags: [viewers]
      summary: Build a declarative ViewSpec for something to view (renderer-agnostic; consumed by the viewer launchers)
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - in: path
          name: kind
          required: true
          schema: { type: string, enum: [subject, simulation, analysis, group, custom] }
        - in: query
          name: subject
          required: false
          schema: { type: string }
        - in: query
          name: simulation
          required: false
          schema: { type: string }
        - in: query
          name: space
          required: false
          schema: { type: string, enum: [subject, mni] }
        - in: query
          name: field
          required: false
          schema: { type: string }
        - in: query
          name: analysis
          required: false
          schema: { type: string }
        - in: query
          name: atlas
          required: false
          description: >-
            which atlas overlay to build, as an id from GET /api/catalog/atlases for the same
            subject and space (kind=subject), or the basename of a bundled MNI atlas
            (kind=group, or kind=subject with space=mni). Absent preserves the previous
            behaviour exactly -- the server picks (labeling.nii.gz, else the first listed
            subject atlas; DEFAULT_MNI_ATLAS in MNI). An unknown id falls back to that same
            server choice rather than 404-ing, so a stale bookmark still renders something.
          schema: { type: string }
        - in: query
          name: roi
          required: false
          schema: { type: string }
        - in: query
          name: path
          required: false
          description: container path to an arbitrary volume/mesh, for kind=custom
          schema: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ViewSpec"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: unknown subject/simulation/analysis
      # POST /api/view/args (freeview_args/freeview_command preview + scene) and the removed D3
      # /api/viewers/{freeview,gmsh} launch routes are deliberately not declared here, matching
      # this contract's existing practice of documenting only a subset of the live API (see
      # contracts/SCHEMA-CHANGES.md) -- /api/view/args's request/response shape is exercised by
      # tests/test_catalog_v1.py and desktop/tests/mock-server/server.mjs instead.

  # ------------------------------------------------------------------------------------- files (v1)
...
  /api/pipelines/validate:
    post:
      tags: [pipelines]
      summary: Validate a pipeline graph, with a reason per problem
      description: >-
        Never 500s on a graph problem: a cycle, an incompatible port or an unbound required input
        comes back as an `issue` with a sentence the UI shows next to the node or wire. A 422 means
        the document is not shaped like a pipeline at all.
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PipelineRequest" }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PipelineValidation" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          description: the document is not a valid pipeline document
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: tit_session
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    Unauthorized:
      description: missing or invalid session/token
...
  schemas:
...
    PipelineKind:
      type: string
      enum:
        [
          pre,
          sim,
          flex,
          flex_adaptive,
          flex_pareto,
          ex,
          mex,
          leadfield,
          analyzer,
          stats,
          source,
          blender,
          nifti_average,
          nilearn,
        ]
    ValidateError:
      type: object
      required: [path, message]
      properties:
        path: { type: string }
        message: { type: string }
    ValidateResult:
      type: object
      required: [ok, errors]
      properties:
        ok: { type: boolean }
        errors:
          type: array
          items: { $ref: "#/components/schemas/ValidateError" }
...
    PipelineNode:
      type: object
      description: One node = exactly one existing job kind carrying exactly the config that page builds.
      required: [id, kind]
      properties:
        id: { type: string, description: canvas-scoped node id, unique within the document }
        kind:
          type: string
          enum: [pre, leadfield, flex, ex, mex, sim, analyzer, source, stats]
        label: { type: string, description: user-visible name; defaults to the kind }
        config:
          $ref: "#/components/schemas/PipelineConfig"
        position:
          type: object
          description: canvas coordinates; carried through untouched by the server
          properties:
            x: { type: number }
            y: { type: number }
    PipelineEdge:
      type: object
      description: >-
        A typed wire. `port` is one named output of `from` bound into the same-named input of `to`
        -- which is what makes "Optimizer -> Simulator" mean "the simulator's montage list is the
        optimizer's result" rather than a drawn line with no semantics.
      required: [from, to, port]
      properties:
        from: { type: string }
        to: { type: string }
        port:
          type: string
          enum: [subjects, montages, simulation, roi, leadfield]
    Pipeline:
      type: object
      description: The canvas document; mirrors contracts/pipeline.schema.json exactly.
      required: [version, nodes]
      properties:
        version: { type: integer, enum: [1] }
        name: { type: string }
        nodes:
          type: array
          items: { $ref: "#/components/schemas/PipelineNode" }
        edges:
          type: array
          items: { $ref: "#/components/schemas/PipelineEdge" }
    PipelineRequest:
      type: object
      description: >-
        A pipeline document, either sent bare or wrapped as `{pipeline: <document>}` (the wrapped
        form is what `run` uses, since it carries submission options alongside).
      properties:
        pipeline: { $ref: "#/components/schemas/Pipeline" }
        version: { type: integer, enum: [1] }
        name: { type: string }
        nodes:
          type: array
          items: { $ref: "#/components/schemas/PipelineNode" }
        edges:
          type: array
          items: { $ref: "#/components/schemas/PipelineEdge" }
...
    PipelineIssue:
      type: object
      required: [level, message]
      properties:
        level: { type: string, enum: [error, warning] }
        message: { type: string, description: one sentence the UI shows verbatim }
        node_id: { type: string }
        edge: { $ref: "#/components/schemas/PipelineEdge" }
...
    PipelineValidation:
      type: object
      required: [ok, issues, order]
      properties:
        ok: { type: boolean, description: false iff any issue has level "error" }
        issues:
          type: array
          items: { $ref: "#/components/schemas/PipelineIssue" }
        order:
          type: array
          description: node ids in dependency-first order; empty when the graph has a cycle
          items: { type: string }
        jobs:
          type: array
          description: what Run would submit; empty when the pipeline does not validate
          items: { $ref: "#/components/schemas/PipelineJobPreview" }
...
    ViewLayer:
      type: object
      required: [path, kind, colormap, opacity, visible]
      properties:
        path: { type: string }
        kind: { type: string, enum: [volume, label] }
        colormap: { type: string }
        opacity: { type: number, minimum: 0, maximum: 1 }
        visible: { type: boolean }
        cal_min: { type: [number, "null"] }
        cal_max: { type: [number, "null"] }
        lut: { type: [string, "null"] }
        percentile:
          oneOf:
            - type: object
              required: [lo, hi]
              properties:
                lo: { type: number, minimum: 0, maximum: 100 }
                hi: { type: number, minimum: 0, maximum: 100 }
            - type: "null"
          description: >-
            Requested percentile window (e.g. {lo: 95, hi: 99.9}), alternative to explicit
            cal_min/cal_max. The server resolves this against the volume's actual voxel-intensity
            distribution at launch time (GET /api/view/{kind} or POST /api/viewers/freeview),
            filling cal_min/cal_max with the resolved absolute values before responding -- a
            layer's cal_min/cal_max are always the final absolute thresholds a viewer applies,
            whether the caller supplied them directly or via percentile.
    ViewSpec:
      type: object
      required: [space, layers, freeview_args]
      properties:
        space: { type: string, enum: [subject, mni] }
        layers:
          type: array
          items: { $ref: "#/components/schemas/ViewLayer" }
          description: >-
            resolved_layers: by the time a ViewSpec leaves the server (from GET /api/view/{kind}
            or is accepted by POST /api/view/args), every layer's cal_min/cal_max is already
            resolved to concrete numbers -- percentile is carried for the client's display
            only (e.g. to preselect a threshold-mode control), never something the client must
            resolve itself.
        freeview_args:
          type: array
          items: { type: string }
          description: >-
            DEPRECATED (D3, dev/notes/v3-docker-streamline-plan.md): the exact argv tail an
            external Freeview process would need (to_freeview_args). Kept for one release; no
            route launches Freeview any more (there is no X11 in this runtime).
        scene:
          type: [object, "null"]
          description: >-
            The real Tetravox ViewSpec v2 document (tit.viewspec.to_tetravox_viewspec) the
            embedded viewer's Engine.load() accepts directly -- datasets addressed by
            origin-relative /api/files/raw/... URLs (both DatasetRef.path and .absPath), layers
            in the engine's own vocabulary (VolumeLayer/MeshLayer). Derived from `layers` by a
            pure function, so the two always describe the same files. Untyped here (a generic
            object) because the full engine scene model is Tetravox-owned; the subset this
            server emits is hand-schema'd in contracts/tetravox-viewspec-v2.schema.json
            (host-facing). Optional, so an older server still satisfies this contract.

    # ------------------------------------------------------------------------------------ files (v1)
    # (no dedicated components; responses are raw bytes / text / TableData, typed inline above)

    # -------------------------------------------------------------------------------- settings (v1)
...
    SimulationConfig:
      type: object
      additionalProperties: true
      x-tit-config: SimulationConfig
      description: "injected from contracts/schema.json by dev/build_contract.py"
    Montage:
      type: object
      additionalProperties: true
      x-tit-config: Montage
      description: "injected from contracts/schema.json by dev/build_contract.py"
    FlexConfig:
      type: object
      additionalProperties: true
      x-tit-config: FlexConfig
      description: "injected from contracts/schema.json by dev/build_contract.py"
    ExConfig:
      type: object
      additionalProperties: true
      x-tit-config: ExConfig
      description: "injected from contracts/schema.json by dev/build_contract.py"
    MExConfig:
      type: object
      additionalProperties: true
      x-tit-config: MExConfig
      description: "injected from contracts/schema.json by dev/build_contract.py"
    AnalyzerConfig:
      type: object
      additionalProperties: true
      x-tit-config: AnalyzerConfig
      description: "injected from contracts/schema.json by dev/build_contract.py"
    PreprocessConfig:
      type: object
      additionalProperties: true
      x-tit-config: PreprocessConfig
      description: "injected from contracts/schema.json by dev/build_contract.py"
    QSIPrepConfig:
      type: object
      additionalProperties: true
      x-tit-config: QSIPrepConfig
      description: "injected from contracts/schema.json by dev/build_contract.py"
    QSIReconConfig:
      type: object
      additionalProperties: true
      x-tit-config: QSIReconConfig
      description: "injected from contracts/schema.json by dev/build_contract.py"
    GroupComparisonConfig:
      type: object
      additionalProperties: true
      x-tit-config: GroupComparisonConfig
      description: "injected from contracts/schema.json by dev/build_contract.py"
    CorrelationConfig:
      type: object
      additionalProperties: true
      x-tit-config: CorrelationConfig
      description: "injected from contracts/schema.json by dev/build_contract.py"
    SourceConfig:
      type: object
      additionalProperties: true
      x-tit-config: SourceConfig
      description: "injected from contracts/schema.json by dev/build_contract.py"
    LeadfieldConfig:
      type: object
      additionalProperties: true
      x-tit-config: LeadfieldConfig
      description: "injected from contracts/schema.json by dev/build_contract.py"
    MontageConfig:
      type: object
      additionalProperties: true
      x-tit-config: MontageConfig
      description: "injected from contracts/schema.json by dev/build_contract.py"
    VectorConfig:
      type: object
      additionalProperties: true
      x-tit-config: VectorConfig
      description: "injected from contracts/schema.json by dev/build_contract.py"
    RegionConfig:
      type: object
      additionalProperties: true
      x-tit-config: RegionConfig
      description: "injected from contracts/schema.json by dev/build_contract.py"
    NiftiAverageConfig:
      type: object
      additionalProperties: true
      x-tit-config: NiftiAverageConfig
      description: "injected from contracts/schema.json by dev/build_contract.py"
    NilearnConfig:
      type: object
      additionalProperties: true
      x-tit-config: NilearnConfig
      description: "injected from contracts/schema.json by dev/build_contract.py"
    PipelineConfig:
      description: >-
        Union of every config dataclass a job/validate/plan "config" body can carry. The real
        discriminant is the `kind`/path parameter alongside it, not a field on the object itself;
        this union exists so generated TS types cover every shape before dev/build_contract.py
        replaces each placeholder member with its generated schema.
      oneOf:
        - $ref: "#/components/schemas/SimulationConfig"
        - $ref: "#/components/schemas/Montage"
        - $ref: "#/components/schemas/FlexConfig"
        - $ref: "#/components/schemas/ExConfig"
        - $ref: "#/components/schemas/MExConfig"
        - $ref: "#/components/schemas/AnalyzerConfig"
        - $ref: "#/components/schemas/PreprocessConfig"
        - $ref: "#/components/schemas/QSIPrepConfig"
        - $ref: "#/components/schemas/QSIReconConfig"
        - $ref: "#/components/schemas/GroupComparisonConfig"
        - $ref: "#/components/schemas/CorrelationConfig"
        - $ref: "#/components/schemas/SourceConfig"
        - $ref: "#/components/schemas/LeadfieldConfig"
        - $ref: "#/components/schemas/MontageConfig"
        - $ref: "#/components/schemas/VectorConfig"
        - $ref: "#/components/schemas/RegionConfig"
        - $ref: "#/components/schemas/NiftiAverageConfig"
        - $ref: "#/components/schemas/NilearnConfig"